Extern variables were tested. Some bugs fixed.

+++ data_types/atomfloatdata.cpp
@@ -6,6 +6,6 @@ bool AtomFloatData::isEqual(const Data * data) const
     if (data->getDataType() == ATOM_FLOAT)
         return ((AtomFloatData *)data)->num == num;
     if (data->getDataType() == ATOM_INT)
-        return (float)((AtomIntData *)data)->num == num;
+        return (double)((AtomIntData *)data)->num == num;
     return false;
 }
