+++ data_types/atomfloatdata.h
@@ -3,19 +3,19 @@
 
 #include "data.h"
 
-///@brief Represents float
+///@brief Represents double
 class AtomFloatData:public Data
 {
     friend class AtomIntData;
 public:
-    AtomFloatData(float num){this->num = num;}
+    AtomFloatData(double num){this->num = num;}
     std::string toString() const {return std::to_string(num);}
     DataType getDataType() const {return ATOM_FLOAT;}
     Data * getClone() const {return new AtomFloatData(*this);}
     bool isEqual(const Data * data) const;
-    float getNum() const {return num;}
+    double getNum() const {return num;}
 private:
-    float num;
+    double num;
 };
 
 #endif // ATOMFLOATDATA_H
